+++ comp2013/src/main/java/comp2013/View/SnakeScenes/GameScene.java
@@ -1,0 +1,4 @@
+package comp2013.View.SnakeScenes;
+
+public class GameScene {
+}